+++ src/prototypes.h
@@ -6,7 +6,7 @@
 #endif
 
 /* append.c */
-int encode_lzhuf P_((FILE *infp, FILE *outfp, size_t size, size_t *original_size_var, size_t *packed_size_var, char *name, char *hdr_method));
+int encode_lzhuf P_((FILE *infp, FILE *outfp, off_t size, off_t *original_size_var, off_t *packed_size_var, char *name, char *hdr_method));
 /* bitio.c */
 void fillbuf P_((int n));
 unsigned short getbits P_((int n));
@@ -30,7 +30,7 @@ unsigned short decode_p_dyn P_((void));
 void output_dyn P_((unsigned int code, unsigned int pos));
 void encode_end_dyn P_((void));
 /* extract.c */
-int decode_lzhuf P_((FILE *infp, FILE *outfp, size_t original_size, size_t packed_size, char *name, int method, size_t *read_sizep));
+int decode_lzhuf P_((FILE *infp, FILE *outfp, off_t original_size, off_t packed_size, char *name, int method, off_t *read_sizep));
 /* header.c */
 int calc_sum P_((void *p, int len));
 void convert_filename P_((char *name, int len, int size, int from_code, int to_code, char *from_delim, char *to_delim, int case_to));
@@ -53,7 +53,7 @@ unsigned short decode_c_st1 P_((void));
 unsigned short decode_p_st1 P_((void));
 void decode_start_st1 P_((void));
 /* indicator.c */
-void start_indicator P_((char *name, size_t size, char *msg, long def_indicator_threshold));
+void start_indicator P_((char *name, off_t size, char *msg, long def_indicator_threshold));
 void put_indicator P_((long int count));
 void finish_indicator2 P_((char *name, char *msg, int pcnt));
 void finish_indicator P_((char *name, char *msg));
@@ -67,7 +67,7 @@ void decode_start_lz5 P_((void));
 /* lhadd.c */
 FILE *append_it P_((char *name, FILE *oafp, FILE *nafp));
 FILE *build_temporary_file P_((void));
-void temporary_to_new_archive_file P_((size_t new_archive_size));
+void temporary_to_new_archive_file P_((off_t new_archive_size));
 void cmd_add P_((void));
 void cmd_delete P_((void));
 /* lharc.c */
@@ -120,25 +120,25 @@ int encode_alloc P_((int method));
 unsigned int encode P_((struct interfacing *interface));
 unsigned int decode P_((struct interfacing *interface));
 /* util.c */
-size_t copyfile P_((FILE *f1, FILE *f2, size_t size, int text_flg, unsigned int *crcp));
-int encode_stored_crc P_((FILE *ifp, FILE *ofp, size_t size, size_t *original_size_var, size_t *write_size_var));
+off_t copyfile P_((FILE *f1, FILE *f2, off_t size, int text_flg, unsigned int *crcp));
+int encode_stored_crc P_((FILE *ifp, FILE *ofp, off_t size, off_t *original_size_var, off_t *write_size_var));
 boolean archive_is_msdos_sfx1 P_((char *name));
 int xsnprintf P_((char *dest, size_t size, char *fmt, ...));
 char *xstrchr P_((const char *s, int c));
 char *xstrrchr P_((const char *s, int c));
-char *xmemchr P_((const char *s, int c, size_t n));
-char *xmemrchr P_((const char *s, int c, size_t n));
+char *xmemchr P_((const char *s, int c, off_t n));
+char *xmemrchr P_((const char *s, int c, off_t n));
 int str_safe_copy P_((char *dst, const char *src, int dstsz));
 
 /* util.c */
 #if !HAVE_MEMMOVE
-void *memmove P_((void *dst, const void *src, size_t cnt));
+void *memmove P_((void *dst, const void *src, off_t cnt));
 #endif
 #if !HAVE_STRDUP
 char *strdup P_((const char *buf));
 #endif
 #if !HAVE_MEMSET
-char *memset P_((char *s, int c, size_t n));
+char *memset P_((char *s, int c, off_t n));
 #endif
 #if !HAVE_STRCASECMP
 int strcasecmp P_((const char *p1, const char *p2));
